fix(scanner): account for fragile edge cases with unedited text
These unedits were performed after a particular token such as `with` was
ill-formed from an edit (e.g. withs), and as such the edit extends
beyond the original length. When unediting, this token is scanned, and
then a semicolon is returned right after due to the tree being in a
fragile state of errors beforehand