Support for string escapes
This PR adds support for string escapes, as proposed in
https://github.com/tree-sitter/tree-sitter-scala/issues/207

I also slightly changed the definition of `interpolated_string` by
qualifying the open quotes with `token.immediate(…)`. Prior to this
change the rule would incorrectly match `foo ""` as an `interpolated_string_expression`

I ran these changes against all of the `.scala` files in
https://github.com/scala/scala3 and https://github.com/scala/scala.

The files in `scala/scala3` that newly have errors are:
* `tests/neg/fEscapes.scala`
* `tests/neg/unicodeEscapes-interpolations.scala`
* `tests/pos/multiLineOps.scala`
* `tests/run/i14164.scala`

The first two are tests containing examples of invalid escape
sequences that are expected to fail. `multiLineOps.scala` contains a
line `send_! "!"` that now parses with an error. Previously this
parsed as an `interpolated_string_expression`, which is also entirely
incorrect, So this error is a result of the adding
`token.immediate('"')` to the definition of `interpolated_string`, and
I do not think the change is a regression. Similarly, `i14164.scala`
contains a multi-line expression that previously incorrectly parsed to
`interpolated_string_expression`, and now parses more correctly,
though with an error.

The files in `scala/scala` that newly have errors are similar. Two
test files with intentionally broken escape sequences, and the same
`multiLineOps.scala`.