Support for string escapes
This PR adds support for string escapes, as proposed in
https://github.com/tree-sitter/tree-sitter-scala/issues/207

In addition to the updated tests, I ran these changes against all of
the .scala files in https://github.com/scala/scala3. Before this
change there were 1,567 that contained at least one parse error, and
after this change there are 1,570. The newly broken files are
[`tests/neg/fEscapes.scala`](https://github.com/scala/scala3/blob/main/tests/neg/fEscapes.scala),
[`tests/neg/t7292-removal.scala`](https://github.com/scala/scala3/blob/main/tests/neg/t7292-removal.scala),
and
[`tests/neg/unicodeEscapes-interpolations.scala`](https://github.com/scala/scala3/blob/main/tests/neg/unicodeEscapes-interpolations.scala). All
of these files are tests that intentionally contain parse errors.