Remove special casing of raw strings
As suggested by @susliko, this commit removes the special handling of
raw strings so that they are treated like other interpolated
strings. This means that tree-sitter will parse escape sequences in
single-line raw strings, even though Scala will not interpret those
characters as escape sequences, but we think this is a worthwhile
tradeoff for simpler implementation.